steer servo with joystick works. changed can data to int8 so we can send negative numbers

+++ node_2/drivers/can_controller.h
@@ -21,7 +21,7 @@ typedef struct can_message_t
 {
 	uint16_t id;
 	char data_length;
-	int32_t data[8];
+	int8_t data[8];
 } CAN_MESSAGE;
 
 uint8_t can_init_def_tx_rx_mb(uint32_t can_br);
